+++ packages/terra-text/tests/jest/Text.test.jsx
@@ -48,6 +48,16 @@ describe('Text', () => {
     expect(text).toMatchSnapshot();
   });
 
+  it('should set inherit-color class if no colorClass is provided', () => {
+    const text = shallow(<Text>Test</Text>);
+    expect(text).toMatchSnapshot();
+  });
+
+  it('should not set inherit-color class is a colorClass is provided', () => {
+    const text = shallow(<Text colorClass="TestClass">Test</Text>);
+    expect(text).toMatchSnapshot();
+  });
+
   it('should support rendering an array of elements as a children', () => {
     /* eslint-disable comma-dangle */
     const text = shallow(
@@ -98,5 +108,11 @@ describe('Text', () => {
     } catch (e) {
       expect(e.message).toContain('The prop `children` is marked as required');
     }
+  });
+
+  it('should forward ref to span', () => {
+    const ref = React.createRef();
+    mount(<React.Fragment><Text id="TestId" ref={ref}>Test</Text></React.Fragment>);
+    expect(ref.current.id).toEqual('TestId');
   });
 });