+++ packages/terra-show-hide/CHANGELOG.md
@@ -2,7 +2,20 @@ ChangeLog
 =========
 
 Unreleased
+----------
+
+2.22.0 - (September 6, 2019)
 ------------------
+### Added
+* Props table markdown file import added to doc site page
+
+2.21.0 - (September 3, 2019)
+------------------
+### Changed
+* Updated the 'keycode-js' version from '^1.0.4' to '^2.0.1'.
+
+2.20.0 - (August 21, 2019)
+--------------------------
 ### Changed
 * Cleaned up imports in examples and test files
 * Import for package version from package.json added to doc page