+++ packages/terra-toggle-button/CHANGELOG.md
@@ -1,8 +1,21 @@
-Changelog
+ChangeLog
 =========
 
 Unreleased
+----------
+
+3.22.0 - (September 6, 2019)
 ------------------
+### Added
+* Props table markdown file import added to doc site page
+
+3.21.0 - (September 3, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
+3.20.0 - (August 21, 2019)
+--------------------------
 ### Changed
 * Cleaned up imports in examples and test files
 * Import for package version from package.json added to doc page